requirements.txt
Flask==3.1.2
svg.path==7.0
gunicorn==23.0.0
